Initial version of the calculator.
This just does some very basic display of the repayment information.
I don't have the actual numbers yet, so I populated the tranche table
with numbers that make it easy to see that the display and math is working.
I expect this script to be incorported into an existing page, though
I may be wrong in which case I'll improve the appearance of what this
generates.