There are now three <script> blocks with different (and documented) purposes.
The first is where an HTML-author is likely to need to make changes.  It
contains the names of elements in the DOM.

The second is constants associated with the bond such as the
expected amount and date of tranche sales.

The third is the actual JS code.

They're separated as the third (and perhaps second) can be placed
into a separate file.  More significantly, different types of authors
are likely to need to change the content within each block.