Changed text and tranche numbers and recomputing now occurs on any change to the home value input.
The new text and tranche numbers are closer to correct, but it is still
a very rough estimate.  I would like to see written information on the
predicted sales.  Absent that, I can get at least some of the information
from the bond attorney's speaking at the Oct 29 presentation at MHS (which
I can rewatch on video).

The "compute" button is still on the page, but now computation occurs
on any change to the value in the "home value" input field.  This feels
*too* sensitive/dynamic to me, but this seems at least somewhat popular.
I can always retune it to change only when a return is pressed or the
field loses focus or some such thing if that is desired.