Annual payment table and the payment graphic are both working.
Note that there are some slight differences in the payment table
computed here from the document being used as a baseline:
MontclairSD_Bond_Referendum_Tax_Impact_Analysis_September_PEC_Update_Tranche_TAXIMPACT[95].
For example, the total payment in years 2027 and 2028 are listed as
$850 and $849 respectively by this appliction while the spreadsheet's
column R lists both as $849.  The spreadsheet is inaccurate, though, as
Rn should be the sum Fn+Jn+Nn.  J13 != J14, so R13 should not equal R14
yet it does in the spreadsheet.

The output of this page contains an artifact of the table of bond
sales (in the <table> with an ID of "bondSales").  Either the function
to generate this (displayTranches()) must be updated to deal with the
new data structure storing the payment information or those HTML artifacts
should be removed.