Remove a piece of dead code applicable only to the now-replaced storage of tranche data.
Note that the function displayTranches() still has a lot of this dead
code, albeit commented out.  I've left that in place as I've yet to rewrite
it for the new storage structure.  The problem is that it would need
information that can no longer be determined from the payment data.  If that
table of bond sales is required, either additional data must be placed within the
program or some additional deterministic relationship between the
sale of each tranche and the repayment of those bonds must be identified.

I am trying to get the latter information (since of course there must
be some mathematical connection between the bonds sold and repaid!)
but it is a bit of an "operator" (or "telephone") game so far.